### 2019-04-14 #1372 @whedon
[PRE REVIEW]: xml2jupyter: Mapping parameters between XML and Jupyter widgets
**Submitting author:** @rheiland (<a href="http://orcid.org/0000-0002-7440-2905">Randy Heiland</a>)
**Repository:** <a href="https://github.com/rheiland/xml2jupyter" target ="_blank">https://github.com/rheiland/xml2jupyter</a>
**Version:** v1.0.0
