### 2019-04-14 #1372 @rheiland
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.01372/joss.01372/10.21105.joss.01372.pdf)

As a first-time JOSS submitter, I'm not exactly sure how this review thread works, but if I'm supposed to reply to this "check article proof", I would just say that it looks acceptable to me. It's just slightly strange with a line break on a quoted <user_parameters> (p. 2).
