### 2019-04-15 #1372 @rheiland
Looks good - just made a minimal change: "shown below" -> "shown in Figure 3" (because the figure wasn't *below*).
