### 2019-04-18 #1372 @lheagy
:wave: Hi @fbergmann, would you be willing to review this submission for the Journal of Open Source Software (JOSS)?

JOSS is a free, open-source, community driven and developer-friendly online journal. The review process is open, and conversation between the authors, reviewers and editors are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository
and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments
and improve their submission until acceptance (or withdrawal, if they feel unable to
satisfy the review).

If you are unable to review at this time, any suggestions you have for alternate reviewers would be appreciated.

Thanks for your consideration!
