### 2019-04-23 #1372 @lheagy
👋 Hi @dangunter, would you be willing to review this submission for JOSS?
