### 2019-04-23 #1408 @rheiland
Article proof is acceptable to me, just slightly strange with Fig 1 appearing in the middle of an XML block and a couple of line breaks in `quoted_strings`.
