### 2019-04-23 #1408 @lheagy
👋 Hi @choldgraf, thanks for being willing to review 🎉: If you haven't yet accepted the JOSS invite, please do so here: https://github.com/openjournals/joss-reviews/invitations. This will let you check off the boxes in the issue thread above. You will also probably want to set yourself as "not watching" this repo so that you don't get flooded with all of the JOSS-review traffic.

As for the review, there is a checklist above to help guide the process. Please feel free to ping me with any questions!
