### 2019-04-24 rheiland/xml2jupyter#1 @choldgraf
Add instructions to install
From looking at the README, I assume that one is meant to use this script directly, instead of installing via pip etc. It would be helpful for people if there were instructions explicitly stating this, since many people will expect there to be an `## Installation` category (even if the instruction just say "run `git clone XXX` and then see the **Usage** section for information on how to run the software".
