### 2019-04-24 rheiland/xml2jupyter#2 @choldgraf
Add some context for what this project is for
It would be helpful if there were more context about the specific scientific context in which this package is useful. Because the name "xml2jupyter" is fairly generic (XML can store all kinds of information, while this package refers to a specific kind of XML), it's important to signal quickly what kind of users it is targeted toward.
