### 2019-04-24 rheiland/xml2jupyter#3 @choldgraf
Add a simpler example
The demo for this project is *really* cool - though it is a bit complex. I stepped through the GUI, but was still a bit unclear how I could integrate the script into my own workflow. For example, the "usage" section says that you don't technically need Jupyter to run this package, and uses the command line in its sample code, whereas the  example utilizes a pre-rendered GUI. I think there should be a minimal example showing the XML structure that this supports, the command to process it, and the kind of output it generates.
