### 2019-04-24 rheiland/xml2jupyter#5 @choldgraf
Improve contributing guidelines
There are currently fairly minimal contributing guidelines in the README - they mention GitHub issues, but nothing really about how folks can get involved with the project and submit their own pull requests, and there aren't any guidelines helping people navigate the codebase itself (e.g. how it's structured, overview of scripts/modules/etc). For example, [here's a contributing file for one of our projects](https://github.com/jupyter/repo2docker/blob/master/CONTRIBUTING.md). You don't necessarily need *that* much information, but some more than just "use github issues" would be helpful.
