### 2019-04-24 rheiland/xml2jupyter#5 @choldgraf
(note, if you don't really want people to contribute development, that's fine too...it's just a bit unclear from the contributing guidelines)
