### 2019-04-24 rheiland/xml2jupyter#3 @choldgraf
Another thing I noticed is that the structure of the python configuration file isn't super clearly explained (I also don't see a CLI API documentation in the README anywhere)
