### 2019-04-24 #1408 @choldgraf
Hey all - I took a pass through the repository and have a few comments and thoughts! I've opened issues for each, will take another pass as things get addressed etc!

I'll update this comment periodically as things change.

# Issues added to repository

## Blockers
- [ ] Installation instructions: https://github.com/rheiland/xml2jupyter/issues/1
- [ ] Improved context / purpose of project: https://github.com/rheiland/xml2jupyter/issues/2
- [ ] More simplified example/usage demo: https://github.com/rheiland/xml2jupyter/issues/3
- [ ] Uncomment test code: https://github.com/rheiland/xml2jupyter/issues/new

## Non-blockers
- [ ] Improvements to contributing guidelines: https://github.com/rheiland/xml2jupyter/issues/5
