### 2019-05-06 7aa544d
add more test of command line params
