+++ README.md
@@ -10,7 +10,9 @@ To see an example application that has been generated with xml2jupyter, click th
 <!-- [![Azure Notebooks](https://notebooks.azure.com/launch.png)](https://notebooks.azure.com/import/gh/randy-heiland/xml2jupyter) -->
 
 ## Overview
-Using a [PhysiCell](http://physicell.mathcancer.org/) XML configuration file, generate a graphical user interface (GUI) consisting of [Jupyter widgets](https://ipywidgets.readthedocs.io/en/stable/index.html). The `xml2jupyter.py` Python script provides a core component of the GUI by generating a panel (Jupyter `Tab`) of customized model parameter widgets. The other Python scripts in this repository are static; all together, they provide a self-contained Jupyter notebook GUI tailored to a custom PhysiCell simulator. While the project is specific to PhysiCell, the basic idea of generating Jupyter widgets can be extended to other text-based configuration files.
+Using a [PhysiCell](http://physicell.mathcancer.org/) XML configuration file, generate a graphical user interface (GUI) consisting of [Jupyter widgets](https://ipywidgets.readthedocs.io/en/stable/index.html). The `xml2jupyter.py` Python script provides a core component of the GUI by generating a panel (Jupyter `Tab`) of custom model parameter widgets. The other Python scripts in this repository are static; all together, they provide a self-contained Jupyter notebook GUI tailored to a custom PhysiCell simulator. While the project is specific to PhysiCell, the basic idea of generating Jupyter widgets can be extended to other text-based configuration files.
+
+We realize the name "xml2jupyter" is pretty generic. We wanted a name that was descriptive, but also concise and, besides, its functionality can (and most likely will) be extended well beyond what is possible in the initial release - to generate custom parameter widgets for a PhysiCell configuration file.
 
 <!--
 If you simply want to try the notebook, without downloading anything, try clicking on this Binder badge [![Binder](https://img.shields.io/badge/PhysiCell-JupyterGUI-E66581.svg)](https://mybinder.org/v2/gh/rheiland/xml2gui/master?filepath=PhysiCell.ipynb) to run it from your browser.
@@ -38,9 +40,42 @@ conda install scipy
 conda install jupyter
 ```
 
-## Sample Usage
+## Simple example
 
-From the root directory of xml2jupyter (we demonstrate using Unix-style shell commands):
+The  `xml2jupyter.py` script assumes the XML input file has a block element `<user_parameters>` that contains children
+elements describing PhysiCell simulation parameters. Each parameter element consists of its name with attributes, defining its data type, units (optional), description (optional), whether the widget should be hidden (optional), and the parameter's default value. The attributes will determine the appearance and behavior of the Jupyter widget. For numeric widgets (the most common type for PhysiCell), xml2jupyter will calculate a delta step size as a function of the default value and this step size will be used by the widget's graphical increment/decrement feature.
+
+To illustrate, consider the following simple XML example, containing each of the four (currently) supported data types (double, int, string, bool) and the various attributes:
+
+```
+<PhysiCell_settings>
+  <user_parameters>
+    <radius type="double" units="micron"
+        description="initial tumor radius">250.0
+    </radius>
+    <threads type="int">8</threads>
+    <color type="string" hidden="true">red</color>
+    <fix_persistence type="bool">True</fix_persistence>
+  </user_parameters>
+</PhysiCell_settings>
+```
+
+When we map this into Jupyter widgets, we obtain the following rendered results: 
+Notice the `color` parameter is not displayed since we specified it should be
+hidden in the XML.
+The name of the other parameters, their values, and attributes, if present, are displayed 
+in rows (as disabled Jupyter button widgets). 
[…]
