### 2019-05-06 rheiland/xml2jupyter#2 @rheiland
I've tried to address some of your concerns. Please see the updated README.
