### 2019-05-06 rheiland/xml2jupyter#3 @rheiland
Integrating into one's own workflow currently assumes that workflow involves a custom PhysiCell simulation. I've tried to emphasize this in the README (as well as the paper). I did provide a minimal example showing the supported XML structure in the README. For my part, I think I've been a bit confused about what needs to be said in a JOSS paper vs. a JOSS repo. Obviously, it doesn't hurt to duplicate, but it does seem a bit redundant.

Not sure I understand the comment about a CLI API doc. Don't think there's enough there for an API, but maybe I'm misunderstanding.
