### 2019-05-06 #1408 @rheiland
Hi @lheagy and apologies for taking so long to reply! Thanks for keeping things moving along and thanks to @choldgraf for his initial review and comments. I've tried to address most of them and welcome more feedback.
