### 2019-05-21 rheiland/xml2jupyter#5 @rheiland
I tried to clarify and provide a bit more guidance in the "Contributions" section.
