### 2019-05-22 rheiland/xml2jupyter#3 @choldgraf
re: the CLI stuff, I think I had mis-interpreted the functionality of the package, so feel free to dismiss that :-)

re: what should go into the JOSS paper vs. the repository, I think the repository will have much more potential viewership than the paper, so any "useful" information should be in the repo documentation IMO
