### 2019-05-23 rheiland/xml2jupyter#3 @choldgraf
Looking through this more, I think the documentation would benefit from some basic information about the inputs and outputs. E.g. many Python modules have a section like

```
Inputs
------
input_parameter : type
  description

Outputs
-------
output_one : type
  description
```

and most CLI modules have help associated with them (e.g. if you used [argparse](https://docs.python.org/3/library/argparse.html) to parse inputs, you could provide help for each input).

Could you add in the README a section that explains what each input to the script is, an example of how it's structured, etc, as well as the script will output? E.g., it seems like it outputs a full python module that's meant to be imported (but can only be done from within the same folder since it won't be in the `PATH`)
