### 2019-05-23 rheiland/xml2jupyter#3 @choldgraf
Another question: I've gotten this to correctly output a widget that shows the configuration values. But I haven't figured out how to make the plot with visualizations etc. The documentation says there are "static files" you can add somewhere, but there's no example or instructions for how to do so. Can you add this into the README?
