### 2019-05-23 #1408 @choldgraf
I took another pass through the README - I think that it is definitely improved!

I still have a few issues that I'd like to see addressed, particularly around clarifying the inputs/outputs of the script, on improving the test coverage, and on clarifying how you can incorporate this into a full GUI visualization.
