+++ PhysiCell_GUI/basics.py
@@ -1,235 +1,0 @@
-# Basics Tab
-
-import os
-from ipywidgets import Layout, Label, Text, Checkbox, Button, HBox, VBox, \
-    FloatText, BoundedIntText, BoundedFloatText, HTMLMath, Dropdown
-
-
-class BasicsTab(object):
-
-    def __init__(self):
-        
-        micron_units = Label('micron')   # use "option m" (Mac, for micro symbol)
-
-        constWidth = '180px'
-        tab_height = '500px'
-        label_domain = Label('Domain (micron):')
-        stepsize = 10
-        self.xmin = FloatText(step=stepsize,
-            # description='$X_{min}$',
-            description='Xmin',
-            layout=Layout(width=constWidth),
-        )
-        self.ymin = FloatText(step=stepsize,
-            description='Ymin',
-            layout=Layout(width=constWidth),
-        )
-        self.zmin = FloatText(step=stepsize,
-            description='Zmin',
-            layout=Layout(width=constWidth),
-        )
-        self.xmax = FloatText(step=stepsize,
-            description='Xmax',
-            layout=Layout(width=constWidth),
-        )
-        self.ymax = FloatText(step=stepsize,
-            description='Ymax',
-            layout=Layout(width=constWidth),
-        )
-        self.zmax = FloatText(step=stepsize,
-            description='Zmax',
-            layout=Layout(width=constWidth),
-        )
-#            description='$Time_{max}$',
-        self.tmax = BoundedFloatText(
-            min=0.,
-            max=100000000,
-            step=stepsize,
-            description='Max Time',
-            layout=Layout(width=constWidth),
-        )
-        self.xdelta = BoundedFloatText(
-            min=1.,
-            description='dx',   # '∆x',  # Mac: opt-j for delta
-            layout=Layout(width=constWidth),
-        )
-        self.ydelta = BoundedFloatText(
-            min=1.,
-            description='dy',
-            layout=Layout(width=constWidth),
-        )
-        self.zdelta = BoundedFloatText(
-            min=1.,
-            description='dz',
-            layout=Layout(width=constWidth),
-        )
-
-        x_row = HBox([self.xmin, self.xmax, self.xdelta])
-        y_row = HBox([self.ymin, self.ymax, self.ydelta])
-        z_row = HBox([self.zmin, self.zmax, self.zdelta])
-
-        self.omp_threads = BoundedIntText(
-            min=1,
-            description='# threads',
-            layout=Layout(width=constWidth),
-        )
-        self.toggle_svg = Checkbox(
-            description='Cells',    # SVG
-            layout=Layout(width='150px') )  # constWidth = '180px'
-        self.svg_interval = BoundedIntText(
-            min=1,
-            max=99999999,   # TODO: set max on all Bounded to avoid unwanted default
-            description='every',
-            layout=Layout(width='160px'),
-        )
-        def toggle_svg_cb(b):
-            if (self.toggle_svg.value):
-                # self.svg_t0.disabled = False 
-                self.svg_interval.disabled = False
-            else:
-                # self.svg_t0.disabled = True
-                self.svg_interval.disabled = True
-            
-        self.toggle_svg.observe(toggle_svg_cb)
-
-        self.toggle_mcds = Checkbox(
-            description='Subtrates',   # Full
-            layout=Layout(width='180px'),
-        )
-        self.mcds_interval = BoundedIntText(
-            min=0,
-            max=99999999,
-            description='every',
-            layout=Layout(width='160px'),
-        )
-        def toggle_mcds_cb(b):
-            if (self.toggle_mcds.value):
[…]
