+++ .travis.yml
@@ -9,6 +9,7 @@ install:
 #  - do a basic test, defaulting to use "config.xml"
 script:
   - cd tests
+# test with default config.xml
   - python ../xml2jupyter.py
   - if [[ `diff user_params.py user_params_baseline_default.py` ]]; then >&2 echo "Fail"; exit 1; fi
 
