+++ tests/user_params_baseline_cancer_immune.py
@@ -1,0 +1,451 @@
+ 
+# This file is auto-generated from a Python script that parses a PhysiCell configuration (.xml) file.
+#
+# Edit at your own risk.
+#
+import os
+from ipywidgets import Label,Text,Checkbox,Button,HBox,VBox,FloatText,IntText,BoundedIntText,BoundedFloatText,Layout,Box
+    
+class UserTab(object):
+
+    def __init__(self):
+        
+        micron_units = Label('micron')   # use "option m" (Mac, for micro symbol)
+
+        constWidth = '180px'
+        tab_height = '500px'
+        stepsize = 10
+
+        #style = {'description_width': '250px'}
+        style = {'description_width': '25%'}
+        layout = {'width': '400px'}
+
+        name_button_layout={'width':'25%'}
+        widget_layout = {'width': '15%'}
+        units_button_layout ={'width':'15%'}
+        desc_button_layout={'width':'45%'}
+
+        param_name1 = Button(description='random_seed', disabled=True, layout=name_button_layout)
+        param_name1.style.button_color = 'lightgreen'
+
+        self.random_seed = IntText(
+          value=0,
+          step=1,
+          style=style, layout=widget_layout)
+
+        param_name2 = Button(description='immune_activation_time', disabled=True, layout=name_button_layout)
+        param_name2.style.button_color = 'tan'
+
+        self.immune_activation_time = FloatText(
+          value=20160,
+          step=1000,
+          style=style, layout=widget_layout)
+
+        param_name3 = Button(description='save_interval_after_therapy_start', disabled=True, layout=name_button_layout)
+        param_name3.style.button_color = 'lightgreen'
+
+        self.save_interval_after_therapy_start = FloatText(
+          value=3.0,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name4 = Button(description='immune_o2_relative_uptake', disabled=True, layout=name_button_layout)
+        param_name4.style.button_color = 'tan'
+
+        self.immune_o2_relative_uptake = FloatText(
+          value=0.1,
+          step=0.01,
+          style=style, layout=widget_layout)
+
+        param_name5 = Button(description='immune_apoptosis_rate', disabled=True, layout=name_button_layout)
+        param_name5.style.button_color = 'lightgreen'
+
+        self.immune_apoptosis_rate = FloatText(
+          value=6.944e-5,
+          step=1e-05,
+          style=style, layout=widget_layout)
+
+        param_name6 = Button(description='immune_motility_persistence_time', disabled=True, layout=name_button_layout)
+        param_name6.style.button_color = 'tan'
+
+        self.immune_motility_persistence_time = FloatText(
+          value=10,
+          step=1,
+          style=style, layout=widget_layout)
+
+        param_name7 = Button(description='immune_migration_speed', disabled=True, layout=name_button_layout)
+        param_name7.style.button_color = 'lightgreen'
+
+        self.immune_migration_speed = FloatText(
+          value=1.0,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name8 = Button(description='immune_migration_bias', disabled=True, layout=name_button_layout)
+        param_name8.style.button_color = 'tan'
+
+        self.immune_migration_bias = FloatText(
+          value=0.5,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name9 = Button(description='immune_relative_adhesion', disabled=True, layout=name_button_layout)
+        param_name9.style.button_color = 'lightgreen'
+
+        self.immune_relative_adhesion = FloatText(
+          value=0,
+          step=0.01,
+          style=style, layout=widget_layout)
+
+        param_name10 = Button(description='immune_relative_repulsion', disabled=True, layout=name_button_layout)
[…]
