+++ tests/README.md
@@ -1,0 +1,4 @@
+# Automated testing in TravisCI
+
+This directory is used for our automated testing and should probably be considered off-limits to xml2jupyter users.
+
