+++ tests/user_params_baseline_cancer_biobots_red_green.py
@@ -1,0 +1,312 @@
+ 
+# This file is auto-generated from a Python script that parses a PhysiCell configuration (.xml) file.
+#
+# Edit at your own risk.
+#
+import os
+from ipywidgets import Label,Text,Checkbox,Button,HBox,VBox,FloatText,IntText,BoundedIntText,BoundedFloatText,Layout,Box
+    
+class UserTab(object):
+
+    def __init__(self):
+        
+        micron_units = Label('micron')   # use "option m" (Mac, for micro symbol)
+
+        constWidth = '180px'
+        tab_height = '500px'
+        stepsize = 10
+
+        #style = {'description_width': '250px'}
+        style = {'description_width': '25%'}
+        layout = {'width': '400px'}
+
+        name_button_layout={'width':'25%'}
+        widget_layout = {'width': '15%'}
+        units_button_layout ={'width':'15%'}
+        desc_button_layout={'width':'45%'}
+
+        param_name1 = Button(description='cargo_signal_D', disabled=True, layout=name_button_layout)
+        param_name1.style.button_color = 'red'
+
+        self.cargo_signal_D = FloatText(
+          value=1e3,
+          step=100,
+          style=style, layout=widget_layout)
+
+        param_name2 = Button(description='cargo_signal_decay', disabled=True, layout=name_button_layout)
+        param_name2.style.button_color = 'green'
+
+        self.cargo_signal_decay = FloatText(
+          value=.4,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name3 = Button(description='director_signal_D', disabled=True, layout=name_button_layout)
+        param_name3.style.button_color = 'red'
+
+        self.director_signal_D = FloatText(
+          value=1e3,
+          step=100,
+          style=style, layout=widget_layout)
+
+        param_name4 = Button(description='director_signal_decay', disabled=True, layout=name_button_layout)
+        param_name4.style.button_color = 'green'
+
+        self.director_signal_decay = FloatText(
+          value=.1,
+          step=0.01,
+          style=style, layout=widget_layout)
+
+        param_name5 = Button(description='elastic_coefficient', disabled=True, layout=name_button_layout)
+        param_name5.style.button_color = 'red'
+
+        self.elastic_coefficient = FloatText(
+          value=0.05,
+          step=0.01,
+          style=style, layout=widget_layout)
+
+        param_name6 = Button(description='worker_motility_persistence_time', disabled=True, layout=name_button_layout)
+        param_name6.style.button_color = 'green'
+
+        self.worker_motility_persistence_time = FloatText(
+          value=5.0,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name7 = Button(description='worker_migration_speed', disabled=True, layout=name_button_layout)
+        param_name7.style.button_color = 'red'
+
+        self.worker_migration_speed = FloatText(
+          value=5.0,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name8 = Button(description='attached_worker_migration_bias', disabled=True, layout=name_button_layout)
+        param_name8.style.button_color = 'green'
+
+        self.attached_worker_migration_bias = FloatText(
+          value=1.0,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name9 = Button(description='unattached_worker_migration_bias', disabled=True, layout=name_button_layout)
+        param_name9.style.button_color = 'red'
+
+        self.unattached_worker_migration_bias = FloatText(
+          value=0.5,
+          step=0.1,
+          style=style, layout=widget_layout)
+
+        param_name10 = Button(description='number_of_directors', disabled=True, layout=name_button_layout)
+        param_name10.style.button_color = 'green'
+
+        self.number_of_directors = IntText(
+          value=15,
[…]
