+++ tests/mygui_test.py
@@ -1,0 +1,196 @@
+# Rf  https://www.python.org/dev/peps/pep-0008/
+import ipywidgets as widgets
+import xml.etree.ElementTree as ET  # https://docs.python.org/2/library/xml.etree.elementtree.html
+import os
+import glob
+import shutil
+import datetime
+import subprocess
+from pathlib import Path
+from config import ConfigTab
+from about import AboutTab
+from user_params import UserTab
+from svg import SVGTab
+from substrates import SubstrateTab
+
+# join_our_list = "(Join/ask questions at https://groups.google.com/forum/#!forum/physicell-users)\n"
+
+debug_view = widgets.Output(layout={'border': '1px solid black'})
+
+run_on_binder = False
+
+constWidth = '180px'
+
+# create the tabs, but don't display yet
+about_tab = AboutTab()
+config_tab = ConfigTab()
+user_tab = UserTab()
+svg = SVGTab()
+sub = SubstrateTab()
+
+main_xml_filename = '../config_samples/config_biorobots.xml'
+full_xml_filename = os.path.abspath(main_xml_filename)
+#print('full_xml_filename=',full_xml_filename)
+
+tree = ET.parse(full_xml_filename)  # this file cannot be overwritten; part of tool distro
+xml_root = tree.getroot()
+
+nanoHUB_flag = False; # True/False (running on nanoHUB or not)
+if( 'HOME' in os.environ.keys() ):
+	nanoHUB_flag = "home/nanohub" in os.environ['HOME']
+
+def read_config_cb(_b):
+    with debug_view:
+        print("read_config", read_config.value)
+        # e.g.  "DEFAULT" -> read_config /Users/heiland/dev/pc4nanobio/data/nanobio_settings.xml
+        #       "<time-stamp>" -> read_config /Users/heiland/.cache/pc4nanobio/pc4nanobio/59c60c0a402e4089b71b140689075f0b
+        #       "t360.xml" -> read_config /Users/heiland/.local/share/pc4nanobio/t360.xml
+
+    if read_config.value is None:  #rwh: happens when a Run just finishes and we update pulldown with the new cache dir??
+        return
+
+    if os.path.isdir(read_config.value):
+        is_dir = True
+#        config_file = os.path.join(read_config.value, 'config.xml')
+        config_file = os.path.join(read_config.value, full_xml_filename)
+    else:
+        is_dir = False
+        config_file = read_config.value
+
+    fill_gui_params(config_file)  #rwh: should verify file exists!
+    
+    # update visualization tabs
+    if is_dir:
+        svg.update(read_config.value)
+        sub.update(read_config.value)
+    else:  # may want to distinguish "DEFAULT" from other saved .xml config files
+        svg.update('')
+        sub.update('')
+
+
+def write_config_file(name):
+    # Read in the default xml config file, just to get a valid 'root' to populate a new one
+#    full_xml_filename = os.path.abspath('config.xml')
+#    full_xml_filename = os.path.abspath(full_xml_filename)
+    tree = ET.parse(full_xml_filename)  # this file cannot be overwritten; part of tool distro
+    xml_root = tree.getroot()
+    config_tab.fill_xml(xml_root)
+    user_tab.fill_xml(xml_root)
+    tree.write(name)
+
+def get_config_files():
+#    cf = {'DEFAULT': os.path.abspath('data/nanobio_settings.xml')}
+#    cf = {'DEFAULT': os.path.abspath('config.xml')}
+    cf = {'DEFAULT': full_xml_filename}
+    dirname = os.path.expanduser('~/.local/share/pc4nanobio')
+    try:
+        os.makedirs(dirname)
+    except:
+        pass
+    files = glob.glob("%s/*.xml" % dirname)
[…]
