+++ README.md
@@ -106,7 +106,32 @@ If you don't have your current working directory in your PATH, you will need to 
 
 ## PhysiCell simulation GUI
 
-The `xml2jupyter.py` script generates a single panel of widgets that represent custom parameters for a PhysiCell simulation. But our ultimate goal is to provide a full-featured GUI for a PhysiCell simulation. By providing additional, static Python modules to accompany the dynamically-generated custom parameters module (from `xml2jupyter.py`), we can achieve this. The following image shows various tabs in a Jupyter notebook GUI for a PhysiCell simulator (that happens to be running on nanoHUB).
+The `xml2jupyter.py` script generates a single panel of widgets that represent custom parameters for a PhysiCell simulation. But our ultimate goal is to provide a full-featured GUI for a PhysiCell simulation. By providing additional, static Python modules to accompany the dynamically-generated custom parameters module (from `xml2jupyter.py`), we can achieve this.
+To demonstrate, we provide an example GUI for PhysiCell's `cancer biorobots` model, with sample data, in 
+this repository. (Note that you won't be able to "Run" an actual simulation though, without additional effort).
+From the `PhysiCell_GUI` directory, run the `mygui.ipynb` notebook:
+```
+$ cd PhysiCell_GUI
+$ jupyter notebook mygui.ipynb
+```
+The notebook should appear in your browser and look like this:
+
+![mygui startup](paper/images/mygui_startup.png)
+
+Clicking "Run All" in the "Cell" menu should display the full GUI, with the `About` tab's content displayed:
+
+![model description](paper/images/mygui_about.png)
+
+You can click on the other tabs to see their content, e.g., the `User Params` tab which was generated by `xml2jupyter.py`:
+
+![custom user parameters](paper/images/mygui_user_params.png)
+
+and the `Cell Plots` and `Substrate Plots` which will display pre-computed data when you click the `frame` slider:
+
+![cells](paper/images/mygui_cells.png)
+![substrate fields](paper/images/mygui_substrates.png)
+
+The following images show a Jupyter notebook GUI for the same model, with additional substrate plots, running on [nanoHUB](https://nanohub.org/).
 
 ![GUI tabs for PhysiCell simulator.](paper/images/nanohub_cancerbots_2x2.png)
 
