### 2019-05-24 3277332
describe modules for full GUI
