+++ README.md
@@ -106,6 +106,16 @@ If you don't have your current working directory in your PATH, you will need to 
 
 ## PhysiCell simulation GUI
 
+The `PhysiCell_GUI` subdirectory contains Python modules that are responsible for creating and operating the full Jupyter notebook GUI:
+```
+mygui.py  - top-level module that imports the other modules (that create the "tabs") and populates their content
+about.py  - creates the "About" tab with a model's documentation
+config.py	- creates the "Config Basics" tab with general simulation parameters
+user_params.py - creates the "User Params" tab with a user's custom simulation parameters (this module is generated by xml2jupyter.py)
+svg.py    - creates the "Out: Cell Plots" tab to visualize cells (data happens to be in SVG format)
+substrates.py  - creates the "Out: Substrate Plots" tab to visualize substrates 
+```
+
 The `xml2jupyter.py` script generates a single panel of widgets that represent custom parameters for a PhysiCell simulation. But our ultimate goal is to provide a full-featured GUI for a PhysiCell simulation. By providing additional, static Python modules to accompany the dynamically-generated custom parameters module (from `xml2jupyter.py`), we can achieve this.
 To demonstrate, we provide an example GUI for PhysiCell's `cancer biorobots` model, with sample data, in 
 this repository. (Note that you won't be able to "Run" an actual simulation though, without additional effort).
