### 2019-05-24 rheiland/xml2jupyter#3 @rheiland
Added documentation header to xml2jupyter.py. Still not using argparse, but familiar with it and may reconsider in the future. Also added brief descriptions of modules in README.

Added section in README describing how one runs the full GUI (with the sample data visualized).
