### 2019-06-19 10a92dd
add usage information about inputs/outputs to readme
