### 2019-06-19 rheiland/xml2jupyter#3 @choldgraf
Nice - that definitely helps. I added that content to the README in a pull request here: https://github.com/rheiland/xml2jupyter/pull/6 I think that would make it more discoverable
