### 2019-06-19 #1408 @choldgraf
I took a look at the latest changes to the repository, and I think that it's much improved. There's more information in there now about the reasoning behind the package, and how to use it. I consolidated some of the material that was in the python script header as well as the JOSS paper and placed it in the README in order to make it discoverable (https://github.com/rheiland/xml2jupyter/pull/6). I think that beyond this, my comments have been addressed!
