### 2019-06-20 #1408 @rheiland
Thank you both! I've merged Chris's PR and replied to his suggestion on looking into pytest.
