### 2019-06-20 #1408 @lheagy
Excellent, many thanks @choldgraf for your comments and nice work addressing these comments @rheiland! @choldgraf: there are a few unchecked items on the checklist above - if those have all been addressed, would you mind ticking them off, or if there are items that still need to be taken care of, please let us know. Thank you both!
