### 2019-06-20 #1408 @choldgraf
a couple quick questions:

1. It asks to list authors, there is the primary author in the metadata of the paper.md file, so I'm assuming that's a ✔️
2. regarding the version, this repository isn't installable via a package manager, it's "just" a python script that is run at the command line, so I'm not sure how to answer this one. There is a release made, so I'm just considering this a ✔️ as well
3. For the references, I believe that the DOI is missing for Thomas Kluyver's article: (doi:10.3233/978-1-61499-649-1-87) as well as for the scipy reference (DOI:10.1109/MCSE.2007.58)
