### 2019-06-21 #1408 @rheiland
@choldgraf, regarding the scipy ref DOI, the DOI you provided is for "Python for Scientific Computing" by Oliphant. Are you saying that's the preferred ref for citing scipy? If so, I'll need to change the entire bib entry. I was trying to follow advice here: https://www.scipy.org/citing.html
