### 2019-06-21 #1408 @choldgraf
Ah, I was just copy pasting the DOIs that I had found next to the references on the scipy docs. But if you've thought it through then I am +1 on whatever you came up with
