### 2019-06-23 f290ced
new ref for scipy, numpy, etc
