+++ paper/paper.bib
@@ -49,6 +49,7 @@
 	Editor = {F. Loizides and B. Schmidt},
 	Organization = {IOS Press},
 	Pages = {87 - 90},
+	doi = {10.3233/978-1-61499-649-1-87},
 	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
 	Year = {2016}
 }
@@ -82,11 +83,23 @@ doi={10.1038/d41586-018-07196-1},
 url={https://www.nature.com/articles/d41586-018-07196-1}
 }
 
-@Misc{scipy,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001},
-  url = "http://www.scipy.org/",
-  note = {[Online; accessed <today>]}
-}
+@article{Oliphant:2007:PSC:1251563.1251830,
+ author = {Oliphant, Travis E.},
+ title = {Python for Scientific Computing},
+ journal = {Computing in Science and Engg.},
+ issue_date = {May 2007},
+ volume = {9},
+ number = {3},
+ month = may,
+ year = {2007},
+ issn = {1521-9615},
+ pages = {10--20},
+ numpages = {11},
+ url = {http://dx.doi.org/10.1109/MCSE.2007.58},
+ doi = {10.1109/MCSE.2007.58},
+ acmid = {1251830},
+ publisher = {IEEE Educational Activities Department},
+ address = {Piscataway, NJ, USA},
+ keywords = {Python, Python, computer languages, scientific programming, scientific computing, computer languages, scientific computing, scientific programming},
+} 
 
