### 2019-06-23 6f3b8af
change scipy ref
