### 2019-06-23 #1408 @rheiland
Giving this a bit more consideration, I decided to use the Oliphant 2007 citation. SciPy requires NumPy after all, which is covered in the 2007 article; it also gives deserved credit to Oliphant. Thanks for your attention to detail on these DOIs.
