### 2019-06-24 #1408 @lheagy
:wave: @arfon (sorry for asking this again, I know you told me the solution, but I couldn't find the issue where we corrected this elsewhere!): Is there a way to keep a block of code together on one page? Is it something along the lines of a `\clearpage` in the paper? Thanks for your help!
