### 2019-06-24 #1408 @arfon
I think you can do `\newpage` ?
