### 2019-06-24 #1408 @rheiland
@lheagy If you're saying to literally add that command to the paper.md (before the XML code block), I did so, and pushed it. But I confess I don't understand the magic you use to generate the pdf from the .md. As you can see from https://github.com/rheiland/xml2jupyter/blob/master/paper/gen_pdf.sh , I came up with my own hack to generate a .pdf, which didn't split the XML block across pages, fwiw:
https://github.com/rheiland/xml2jupyter/blob/master/paper/paper.pdf
