### 2019-06-24 #1408 @arfon
@rheiland you can run `@whedon generate pdf` yourself too :-) The command we're running is a bit of a monster: https://github.com/openjournals/whedon/blob/c34e1031c081b4e24cd7141023ed81b643c04364/lib/whedon/compilers.rb#L126-L149
