### 2019-06-24 #1408 @rheiland
@arfon Thanks for pointing out the fact that I could have whedon generate the pdf myself (apologies for not remembering that when I listed whedon's commands a long time ago).

Some things that are less than ideal in the pdf, but not sure what, if anything, can be done about them. And besides, they are not critical.

* have Figure 1 appear *before* the associated XML block (and the paragraph that mentions it).

* have a filename be split across lines in the pdf:
```
The xml2jupyter.py script parses the XML and generates a Python module, user_param
s.py, ...
```
