### 2019-06-25 #1408 @arfon
> - have Figure 1 appear before the associated XML block (and the paragraph that mentions it).

Just to confirm, do you want the XML before the figure or vice versa? The current proof looks like this to me:
