### 2019-06-25 #1408 @rheiland
The XML before the figure would be more logical.

On Tue, Jun 25, 2019 at 8:04 AM Arfon Smith <notifications@github.com>
wrote:

>
>    - have Figure 1 appear before the associated XML block (and the
>    paragraph that mentions it).
>
> Just to confirm, do you want the XML before the figure or vice versa? The
> current proof looks like this to me:
>
> [image: Screen Shot 2019-06-25 at 8 03 29 AM]
> <https://user-images.githubusercontent.com/4483/60096768-cff17900-971f-11e9-98d4-0a231868f18e.png>
>
