### 2019-06-25 #1408 @arfon
@rheiland - I don't think Whedon knows how to build a change from my pull request (https://github.com/rheiland/xml2jupyter/pull/7) but when I build the PDF locally with the changes in https://github.com/rheiland/xml2jupyter/pull/7 I get this:

[10.21105.joss.01408.pdf](https://github.com/openjournals/joss-reviews/files/3325993/10.21105.joss.01408.pdf)
