### 2019-06-25 #1408 @rheiland
Beautiful, IMO :-)  Thanks for your PR! In fact, the pdf now looks *very*
similar to the one I generated using my hacky script.

On Tue, Jun 25, 2019 at 11:28 AM Arfon Smith <notifications@github.com>
wrote:

> How's that looking @rheiland <https://github.com/rheiland>?
>
