### 2019-06-25 #1408 @lheagy
Excellent, thanks for your help @arfon 🎉

@rheiland: can you please generate a new release of  `xml2jupyter`, archive it on [zenodo](http://zenodo.org) or similar and post the doi here. Please make sure that the author list and title match the those on the paper.
