### 2019-06-25 #1408 @rheiland
@lheagy: will it be OK to leave my gen_pdf.sh script (and a resulting .pdf) in the /paper directory when I generate a new release, or would that cause problems/confusion?
I also have a /preprint directory that I will remove before doing a new release.
