### 2019-06-25 #1408 @lheagy
@rheiland: I think leaving the `gen_pdf.sh` is fine, but the pdf might cause confusion (particularly as the DOI is not assigned in the one in your repo). So I would recommend removing it. If you wanted to keep this in your repo, then keeping the `preprint` directory and moving the pdf there would be appropriate.
