+++ preprint/README.md
@@ -1,0 +1,5 @@
+# Creating a preprint
+
+* copy both files to the ../paper directory  (to use /images)
+* may have to run 'bibtex preprint' to resolve refs
+* pdflatex preprint
