+++ preprint/preprint.bib
@@ -1,4 +1,4 @@
-@article {PhysiCell:2018,
+@article{PhysiCell:2018,
     author = {Ahmadreza Ghaffarizadeh and Randy Heiland and Samuel H. Friedman and Shannon M. Mumenthaler and Paul Macklin},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
@@ -27,7 +27,7 @@
 
 }
 
-@article {Hunter:2007,
+@Article{Hunter:2007,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
   Journal   = {Computing In Science \& Engineering},
@@ -49,6 +49,7 @@
 	Editor = {F. Loizides and B. Schmidt},
 	Organization = {IOS Press},
 	Pages = {87 - 90},
+	doi = {10.3233/978-1-61499-649-1-87},
 	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
 	Year = {2016}
 }
@@ -71,7 +72,7 @@ doi={10.1515/ntrev-2012-0043},
 url={https://www.degruyter.com/view/j/ntrev.2013.2.issue-1/ntrev-2012-0043/ntrev-2012-0043.xml}
 }
 
-@article {Nature_2018_Jupyter,
+@Article{Nature_2018_Jupyter,
 author={Jeffrey M. Perkel},
 title={Why Jupyter is data scientists’ computational notebook of choice},
 year={2018},
@@ -82,10 +83,23 @@ doi={10.1038/d41586-018-07196-1},
 url={https://www.nature.com/articles/d41586-018-07196-1}
 }
 
-@misc {Jones:2001,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001},
-  url = "http://www.scipy.org/",
-  note = {[Online; accessed April 4, 2019]}
-}
+@article{Oliphant:2007:PSC:1251563.1251830,
+ author = {Oliphant, Travis E.},
+ title = {Python for Scientific Computing},
+ journal = {Computing In Science \& Engineering},
+ issue_date = {May 2007},
+ volume = {9},
+ number = {3},
+ month = may,
+ year = {2007},
+ issn = {1521-9615},
+ pages = {10--20},
+ numpages = {11},
+ url = {http://dx.doi.org/10.1109/MCSE.2007.58},
+ doi = {10.1109/MCSE.2007.58},
+ acmid = {1251830},
+ publisher = {IEEE Educational Activities Department},
+ address = {Piscataway, NJ, USA},
+ keywords = {Python, Python, computer languages, scientific programming, scientific computing, computer languages, scientific computing, scientific programming},
+} 
+
