+++ preprint/preprint.tex
@@ -1,5 +1,6 @@
 \documentclass[10pt,letterpaper]{article}
-\usepackage[top=0.85in,left=2.75in,footskip=0.75in,marginparwidth=2in]{geometry}
+% \usepackage[top=0.85in,left=2.75in,footskip=0.75in,marginparwidth=2in]{geometry}
+\usepackage[top=0.85in,left=1.25in,footskip=0.75in,marginparwidth=2in]{geometry}
 
 % use Unicode characters - try changing the option if you run into troubles with special characters (e.g. umlauts)
 \usepackage[utf8]{inputenc}
@@ -23,7 +24,8 @@
 % text layout - change as needed
 %\raggedright
 \setlength{\parindent}{0.5cm}
-\textwidth 5.25in 
+% \textwidth 5.25in 
+\textwidth 5.5in 
 \textheight 8.75in
 
 % Remove % for double line spacing
@@ -71,7 +73,7 @@
 \reversemarginpar
 
 %%\usepackage{biblatex}
-%%\addbibresource{preprint.bib}
+%%\addbibresource{preprint2.bib}
 
 % document begins here
 \begin{document}
@@ -99,7 +101,9 @@ Paul Macklin\textsuperscript{1,*}
 
 \end{flushleft}
 
-\section*{Abstract}
+\hypertarget{summary}{%
+\section{Summary}\label{summary}}
+
 Jupyter Notebooks \cite{Kluyver:2016aa,Nature_2018_Jupyter} provide
 executable documents (in a variety of programming languages) that can be
 run in a web browser. When a notebook contains graphical widgets, it
@@ -109,24 +113,22 @@ parameter values and run at the command line without a graphical
 interface. Manually editing these files to explore how different values
 affect a simulation can be burdensome for technical users, and
 impossible to use for those with other scientific backgrounds.
-xml2jupyter\footnote{https://github.com/rheiland/xml2jupyter} is a 
-Python package that addresses these scientific
+xml2jupyter\footnote{https://github.com/rheiland/xml2jupyter} is a Python package that addresses these scientific
 bottlenecks. It provides a mapping between configuration files,
 formatted in the Extensible Markup Language (XML), and Jupyter widgets.
 Widgets are automatically generated from the XML file and these can,
 optionally, be incorporated into a larger GUI for a simulation package,
 and optionally hosted on cloud resources. Users modify parameter values
-via the widgets,
-and the values are written to the XML configuration file which is input
-to the simulation's command-line interface. xml2jupyter has been tested
-using PhysiCell \cite{PhysiCell:2018}, an open source, agent-based
-simulator for biology, and it is being used by students for classroom
-and research projects. In addition, we use xml2jupyter to help create
-Jupyter GUIs for PhysiCell-related applications running on nanoHUB \cite{nanoHUB_2013}.
+via the widgets, and the values are written to the XML configuration
+file which is input to the simulation's command-line interface.
+xml2jupyter has been tested using PhysiCell \cite{PhysiCell:2018},
+an open source, agent-based simulator for biology, and it is being used
+by students for classroom and research projects. In addition, we use
+xml2jupyter to help create Jupyter GUIs for PhysiCell-related
+applications running on nanoHUB \cite{nanoHUB_2013}.
 
-\section*{Introduction}
 A PhysiCell configuration file defines model-specific
-\texttt{\textless{}user\char`_parameters\textgreater{}} in XML. Each
+\texttt{\textless{}user\_parameters\textgreater{}} in XML. Each
 parameter element consists of its name with attributes, defining its
 data \emph{type}, \emph{units} (optional), \emph{description}
 (optional), whether the widget should be \emph{hidden} (optional), and
[…]
