+++ paper/paper.md
@@ -100,7 +100,7 @@ $ cp config_samples/config_biorobots.xml .
 $ python xml2jupyter.py config_biorobots.xml
 ```
 The `xml2jupyter.py` script parses the XML and generates a Python
-module named `user_params.py` that contains the Jupyter widgets, together with methods to populate
+module, `user_params.py`, containing the Jupyter widgets, together with methods to populate
 their values from the XML and write their values back to the XML.
 To "validate" the widgets were generated correctly, one could, minimally, open `user_params.py`
 in an editor and inspect it. But to actually see the widgets rendered in a notebook, we provide a simple test:
