### 2019-06-28 #1408 @rheiland
@lheagy @arfon if there's anything that can be done to prevent the line break of a filename (below) in the pdf generation, I'd welcome it. Otherwise, we'll move ahead with generating a release and getting a DOI. Thanks!

![Screen Shot 2019-06-28 at 9 19 06 AM](https://user-images.githubusercontent.com/1774917/60346119-79d33e80-9988-11e9-94b7-77a933b5a3c9.png)
