### 2019-06-28 #1408 @arfon
> @lheagy @arfon if there's anything that can be done to prevent the line break of a filename (below) in the pdf generation, I'd welcome it. Otherwise, we'll move ahead with generating a release and getting a DOI. Thanks!

I'm afraid not. The only thing I can suggest at this point is to slightly adjust the language (lengthen or shorten) to try and force a formatting change.
