### 2019-06-28 #1408 @rheiland
@arfon OK, no worries. I did play a bit with wording, but still ended up with name split across lines, so let's go with what's there.
@lheagy I'm still new to this Zenodo-GitHub-DOI workflow, so I'd welcome feedback if I've mangled it. I did get a DOI that supposedly goes with the (JOSS) Release of our repo - it's 10.5281/zenodo.3261541 .

![Screen Shot 2019-06-28 at 2 42 46 PM](https://user-images.githubusercontent.com/1774917/60364053-3510cd00-99b3-11e9-9d6e-520b4a4b7be8.png)

Is there more I need to do?
