### 2019-06-29 #1408 @MathCancer
Thanks so much to @arfon and @lheagy for all your help in guiding us through the review process! We're okay to let the line break issue go: very much at the point of rapidly diminishing returns, and the proof looks fantastic otherwise!

I noticed an unchecked box on references, but so far as I can tell, all references have a DOI. Is there something else we should fix there?

Thank you! - Paul
