### 2019-06-30 #1408 @lheagy
:wave: @rheiland: would you mind updating the title of the Zenodo archive to be: "xml2jupyter: Mapping parameters between XML and Jupyter widgets" (the same as your paper title)? Please also add all of the authors that are listed on the paper to the archive record: Randy Heiland, Daniel Mishler, Tyler Zhang, Eric Bower, and Paul Macklin.

Thanks!
