### 2019-07-01 #1408 @rheiland
Thank you very much @arfon, @lheagy, and esp @choldgraf for this terrific experience!

Randy
