### 2019-07-01 #1408 @MathCancer
Thanks very much for leading us through our first JOSS submission! Thanks
to @rheiland for terrific work as lead author!

In case it helps your statistics for impact, etc., please note that this
paper has three undergraduate co-authors.

All the best -- Paul

>
